[change] Themeクラスを別ファイルに切り出しました。
