+++ data/src/main/java/net/pantasystem/milktea/data/infrastructure/settings/Theme.kt
@@ -1,0 +1,23 @@
+package net.pantasystem.milktea.data.infrastructure.settings
+
+import net.pantasystem.milktea.model.setting.DefaultConfig
+import net.pantasystem.milktea.model.setting.Theme
+
+fun Theme.toInt(): Int {
+    return when (this) {
+        is Theme.White -> 0
+        is Theme.Black -> 1
+        is Theme.Dark -> 2
+        is Theme.Bread -> 3
+    }
+}
+
+fun Theme.Companion.from(n: Int): Theme {
+    return when (n) {
+        0 -> Theme.White
+        1 -> Theme.Black
+        2 -> Theme.Dark
+        3 -> Theme.Bread
+        else -> DefaultConfig.config.theme
+    }
+}
