+++ model/src/main/java/net/pantasystem/milktea/model/setting/Theme.kt
@@ -1,0 +1,9 @@
+package net.pantasystem.milktea.model.setting
+
+sealed interface Theme {
+    object White : Theme
+    object Black : Theme
+    object Dark : Theme
+    object Bread : Theme
+    companion object
+}
